chore: move network structures
